refactor: :truck: split button class into seperate file from controller
